requirements.txt
google-generativeai
# sqlite3 is built-in
# argparse is built-in
